+++ test/test_running.py
@@ -1,7 +1,6 @@
 import json
 import random
 import re
-import shutil
 import string
 import subprocess
 
@@ -10,23 +9,7 @@ import pytest
 
 #pylint:disable=redefined-outer-name,use-dict-literal,missing-timeout,unspecified-encoding,consider-using-with
 
-from utils import CONTAINER_NAME, PROTO, HOST, login
-
-def dojo_run(*args, **kwargs):
-    kwargs.update(stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
-    return subprocess.run(
-        [shutil.which("docker"), "exec", "-i", CONTAINER_NAME, "dojo", *args],
-        check=kwargs.pop("check", True), **kwargs
-    )
-
-
-def workspace_run(cmd, *, user, root=False, **kwargs):
-    args = [ "enter" ]
-    if root:
-        args += [ "-s" ]
-    args += [ user ]
-    return dojo_run(*args, input=cmd, check=True, **kwargs)
-
+from utils import PROTO, HOST, login, dojo_run, workspace_run
 
 def get_flag(user):
     return workspace_run("cat /flag", user=user, root=True).stdout
